Add age binning, move covid clustering into its own file
