+++ config/global_codebook/global_relationships.py
@@ -229,15 +229,17 @@ def neurodegenerative(responses):
 )
 def age_health_status(responses, binning_config):
     nih_age = responses["nih_age"]
+    if nih_age["response_name"] != "text":
+        return
+    
     age = int(nih_age["response_value"])
 
-    excellent = 0.1
-    very_good = 0.2
-    good = 0.2
-    fair = 0.2
-    poor = 0.2
-    dont_know = 0.05
-    prefer_not_to_answer = 0.05
+    
+    excellent = 0
+    very_good = 0
+    good = 0
+    fair = 0
+    poor = 0
 
     for bin in binning_config:
         start_age = bin["start"]
@@ -255,18 +257,14 @@ def age_health_status(responses, binning_config):
             "Very Good",
             "Good",
             "Fair",
-            "Poor",
-            "Don't know",
-            "Prefer not to state"
+            "Poor"
         ],
         weights=[
             excellent,
             very_good,
             good,
             fair,
-            poor,
-            dont_know,
-            prefer_not_to_answer
+            poor
         ],
         k=1
     )[0]
@@ -401,80 +399,3 @@ def diabetes_types(responses):
                 "response_name": "Skip Logic"
             }
         }
-
-@relationship(
-    name="covid_symptom_clustering",
-    dependencies=None,
-    modifies=[
-        "nih_skin_rash",
-        "nih_conjunctivitis",
-        "nih_red_eyes",
-        "nih_high_temp",
-        "nih_no_sympt",
-        "nih_blue_lips",
-        "nih_balance",
-        "nih_slurred_peech",
-        "nih_neuro_shakes",
-        "nih_numb_extremities",
-        "nih_sweating",
-        "nih_seizures",
-        "nih_rash_toes",
-        "nih_cough",
-        "nih_fever_chills",
-        "nih_diff_breath",
-        "nih_headache",
-        "nih_muscle_ache",
-        "nih_olfactory",
-        "nih_fatigue",
-        "nih_nausea_vomiting_diarrhea",
-        "nih_abdom_pain",
-        "nih_throat_congestion_nose",
-        "nih_other_symp",
-        "nih_wheezing",
-        "nih_confusion",
-        "nih_appetite"
-    ]
-)
-def covid_symptom_clustering(responses, clusters_config):
-    covid_freq = clusters_config["covid_freq"]
-    global_symptoms = clusters_config["global_symptoms"]
-    clusters = clusters_config["cluster_symptoms"]
-    for cluster in clusters:
-        cluster["symptoms"] = {
-            **global_symptoms,
-            **cluster["symptoms"]
-        }
-
-    if random() >= covid_freq: return
-    
-    cluster = choices(
-        clusters,
-        weights=[cluster["frequency"] for cluster in clusters],
-        k=1
-    )[0]
-    
-    modified_responses = {}
-    for variable in cluster["symptoms"]:
-        responses = cluster["symptoms"][variable]
-        used_freq = sum(responses.values())
-        no_change_freq = 1 - used_freq
-
-        response_name = choices(
-            [
-                *responses.keys(),
-                None
-            ],
-            weights=[
-                *responses.values(),
-                no_change_freq
-            ],
-            k=1
-        )[0]
-        if response_name is not None:
-            modified_responses[variable] = {
-                "response_name": response_name
-            }
-    
-    return modified_responses
-        
-
